pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gigatrader"
version = "0.1.0"
description = "Gigatrader - automated equities/options platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alpaca-py>=0.30,<1",
  "python-dotenv>=1",
  "pydantic>=2",
  "pandas>=2",
  "pyarrow>=16",
  "pyyaml>=6",
  "rich>=13",
  "typer>=0.12"
]

[project.scripts]
trade = "app.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "core*", "services*", "cli*", "execution*", "risk*", "strategies*", "backtest*", "ui*"]

[tool.ruff]
line-length = 100
extend-exclude = ["artifacts", "build", "dist", "ui", "tools", "notebooks"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["."]
testpaths = ["tests"]
